Modify: Move Media Package out of Main
