+++ app/src/main/java/com/pleiades/pleione/slotgallery/presentation/media/video/VideoActivity.kt
@@ -1,6 +1,5 @@
-package com.pleiades.pleione.slotgallery.presentation.main.media.video
+package com.pleiades.pleione.slotgallery.presentation.media.video
 
-import android.content.Intent
 import android.content.pm.ActivityInfo
 import android.net.Uri
 import android.os.Bundle
@@ -14,19 +13,23 @@ import androidx.core.view.WindowInsetsControllerCompat
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.MediaItem
 import com.google.android.exoplayer2.Player
+import com.google.android.exoplayer2.ui.StyledPlayerView
 import com.pleiades.pleione.slotgallery.Config.Companion.ACTIVITY_CODE_VIDEO
+import com.pleiades.pleione.slotgallery.Config.Companion.INTENT_EXTRA_NAME
+import com.pleiades.pleione.slotgallery.Config.Companion.INTENT_EXTRA_URI
 import com.pleiades.pleione.slotgallery.R
 import com.pleiades.pleione.slotgallery.databinding.ActivityVideoBinding
 import com.pleiades.pleione.slotgallery.presentation.main.MainActivity
 
 
-class ExternalVideoActivity : AppCompatActivity() {
+class VideoActivity : AppCompatActivity() {
     private lateinit var binding: ActivityVideoBinding
 
     private lateinit var toolbarLayoutParams: ViewGroup.MarginLayoutParams
     private var statusBarHeight = 0
 
     private lateinit var exoPlayer: ExoPlayer
+    private lateinit var playerView: StyledPlayerView
     private lateinit var playerViewLayoutParams: ViewGroup.MarginLayoutParams
     private var navigationBarHeight = 0
 
@@ -55,48 +58,45 @@ class ExternalVideoActivity : AppCompatActivity() {
         // set toolbar margin
         setToolbarMargin(true)
 
-        // check intent
-        if (intent.action == Intent.ACTION_VIEW) {
-            // set title from intent
-            title = intent.data.toString()
+        // set title from intent extra
+        title = intent.getStringExtra(INTENT_EXTRA_NAME)
 
-            // initialize uri from intent
-            val uri = Uri.parse(intent.data.toString())
+        // initialize uri from intent extra
+        val uri = Uri.parse(intent.getStringExtra(INTENT_EXTRA_URI))
 
-            // initialize player
-            binding.playerVideo.setOnClickListener { fullVideo() }
+        // initialize player
+        playerView = findViewById(R.id.player_video)
+        playerView.setOnClickListener { fullVideo() }
 
-            // initialize exoplayer
-            exoPlayer = ExoPlayer.Builder(this).build()
-            exoPlayer.addListener(object : Player.Listener {
-                override fun onPlaybackStateChanged(@Player.State state: Int) {
-                    if (state == Player.STATE_ENDED) {
-                        // un full force
-                        isFull = true
-                        fullVideo()
-                    }
+        // initialize exoplayer
+        exoPlayer = ExoPlayer.Builder(this).build()
+        exoPlayer.addListener(object : Player.Listener {
+            override fun onPlaybackStateChanged(@Player.State state: Int) {
+                if (state == Player.STATE_ENDED) {
+                    // un full force
+                    isFull = true
+                    fullVideo()
                 }
-            })
-            exoPlayer.setMediaItem(MediaItem.fromUri(uri))
+            }
+        })
+        exoPlayer.setMediaItem(MediaItem.fromUri(uri))
 
-            // set exoplayer as player
-            binding.playerVideo.player = exoPlayer
+        // set exoplayer as player
+        playerView.player = exoPlayer
 
-            // initialize player view layout params
-            playerViewLayoutParams = binding.playerVideo.layoutParams as ViewGroup.MarginLayoutParams
+        // initialize player view layout params
+        playerViewLayoutParams = playerView.layoutParams as ViewGroup.MarginLayoutParams
 
-            // initialize navigation bar height
-            val navigationBarHeightResId = resources.getIdentifier("navigation_bar_height", "dimen", "android")
-            navigationBarHeight = resources.getDimensionPixelSize(navigationBarHeightResId)
+        // initialize navigation bar height
+        val navigationBarHeightResId = resources.getIdentifier("navigation_bar_height", "dimen", "android")
+        navigationBarHeight = resources.getDimensionPixelSize(navigationBarHeightResId)
 
-            // set player view margin
-            setPlayerViewMargin(true)
+        // set player view margin
+        setPlayerViewMargin(true)
 
-            // play
-            exoPlayer.prepare()
-            exoPlayer.play()
-        } else
-            finish()
+        // play
+        exoPlayer.prepare()
+        exoPlayer.play()
     }
 
     override fun onResume() {
@@ -202,6 +202,6 @@ class ExternalVideoActivity : AppCompatActivity() {
 
     private fun setPlayerViewMargin(apply: Boolean) {
         playerViewLayoutParams.bottomMargin = if (apply) navigationBarHeight else 0
-        binding.playerVideo.layoutParams = playerViewLayoutParams
+        playerView.layoutParams = playerViewLayoutParams
     }
 }
